Simplify caching of desktop entries
This commit simplifies the caching of desktop entries. Previously
spmenu_run would cache each desktop entry into a cache directory, of
course using it's own format and then parses that yet again to add it to
the arrays. This is terribly inefficient and wastes the user's valuable
space.

With this commit, the entries are parsed and directly written to the six
cache files, which spmenu can then grab from in plain text, without any
parsing. Significantly speeds up the caching and saves disk space.