Allow one more item to be displayed
While this means a VERY insignificant part of the spmenu client is cut
off in some cases (such as on my 1366x768 display), it allows one extra
item to be displayed and importantly for aesthetics, doesn't leave an
empty gap. I consider this worth it, but this might cause problems on
other screen resolutions. Please file a bug report in any such cases.